Time Slow and Joystick
moved joystick code from the playercontroller script to its own new JoyStick script. The joystick is now its own prefab as well.

Added a vignette and bloom post processing effect to the time slow down. The bloom is currently a light blue and the vignette a dark grey. The transition takes .2 seconds.